// src/app/models/Game.ts
export interface Game {
    ID?: string;
    title?: string;
    description?: string;
    image?: string;
    author?: string;
    content?: string;
    content_short?: string;
    publisher?: string;
    publisher_date?: string;
    pages?: string;
    language?: string;
    url_details?: string;
    url_download?: string;
    cover?: string;
    thumbnail?: string;
    num_comments?: string;
    categories?: string;
    tags?: string;

    created_at?: Date
}